+++ src/components/ModelProvider.jsx
@@ -22,10 +22,10 @@ import {PageModelManager} from '@adobe/cq-spa-page-model-manager';
 
 /**
  * Wrapper component responsible for synchronizing a child component with a given portion of the page model.
- * The location of the portion of the page model corresponds to the location of the resource in the page and is accessible via the data_path / page_path properties of the component.
+ * The location of the portion of the page model corresponds to the location of the resource in the page and is accessible via the dataPath / pagePath properties of the component.
  * Those properties are then output in the form of data attributes (data-cq-page-path and data-cq-data-path) to allow the editor to understand to which AEM resource this component corresponds.
  *
- * When the model gets updated the wrapped component gets re-rendered with the latest version of the model passed as the cq_model parameter.
+ * When the model gets updated the wrapped component gets re-rendered with the latest version of the model passed as the cqModel parameter.
  * <p>The ModelProvider supports content items as well as child pages</p>
  *
  * @class
@@ -33,9 +33,9 @@ import {PageModelManager} from '@adobe/cq-spa-page-model-manager';
  * @memberOf components
  *
  * @param {{}} props                      - the provided component properties
- * @param {string} props.data_path        - relative path of the current configuration in the overall page model
- * @param {string} props.page_path        - absolute path of the containing page
- * @param {boolean} props.force_reload    - should the cache be ignored
+ * @param {string} props.dataPath        - relative path of the current configuration in the overall page model
+ * @param {string} props.pagePath        - absolute path of the containing page
+ * @param {boolean} props.forceReload    - should the cache be ignored
  */
 class ModelProvider extends Component {
 
@@ -43,9 +43,9 @@ class ModelProvider extends Component {
         super(props);
 
         this.state = {
-            data_path: props && props.data_path ? props.data_path : '',
-            page_path: props && props.page_path ? props.page_path : '',
-            cq_model: props && props.cq_model
+            dataPath: props && props.dataPath ? props.dataPath : '',
+            pagePath: props && props.pagePath ? props.pagePath : '',
+            cqModel: props && props.cqModel
         };
 
         this.updateData();
@@ -58,7 +58,7 @@ class ModelProvider extends Component {
      */
     updateData() {
         const that = this;
-        const path = this.state.data_path || '';
+        const path = this.state.dataPath || '';
 
         // Fetching the latest data for the item at the given path
         this.getData().then(model => {
@@ -69,20 +69,20 @@ class ModelProvider extends Component {
             model[Constants.DATA_PATH_PROP] = path;
 
             that.setState({
-                data_path: path,
-                page_path: that.getPagePath(),
-                cq_model: model
+                dataPath: path,
+                pagePath: that.getPagePath(),
+                cqModel: model
             });
         });
     }
 
     componentDidMount() {
-        PageModelManager.addListener({pagePath: this.state.page_path, dataPath: this.state.data_path, callback: this.updateData.bind(this)});
+        PageModelManager.addListener({pagePath: this.state.pagePath, dataPath: this.state.dataPath, callback: this.updateData.bind(this)});
     }
 
     componentWillUnmount() {
         // Clean up listener
-        PageModelManager.removeListener({pagePath: this.state.page_path, dataPath: this.state.data_path, callback: this.updateData.bind(this)});
+        PageModelManager.removeListener({pagePath: this.state.pagePath, dataPath: this.state.dataPath, callback: this.updateData.bind(this)});
     }
 
     componentDidUpdate() {
@@ -90,17 +90,17 @@ class ModelProvider extends Component {
     }
 
     componentWillReceiveProps(nextProps) {
-        if (nextProps.data_path !== this.props.data_path || nextProps.page_path !== this.props.page_path) {
+        if (nextProps.dataPath !== this.props.dataPath || nextProps.pagePath !== this.props.pagePath) {
             // Path has been updated.
-            const newDataPath = nextProps.data_path || '';
-            const newPagePath = nextProps.page_path || '';
+            const newDataPath = nextProps.dataPath || '';
+            const newPagePath = nextProps.pagePath || '';
             // Remove old listener associated with the old location
-            PageModelManager.removeListener({pagePath: this.state.page_path, dataPath: this.state.data_path, callback: this.updateData.bind(this)});
+            PageModelManager.removeListener({pagePath: this.state.pagePath, dataPath: this.state.dataPath, callback: this.updateData.bind(this)});
             // Add new listener on the new location.
             // We can not use state because it is not updated yet
             PageModelManager.addListener({pagePath: newPagePath, dataPath: newDataPath, callback: this.updateData.bind(this)});
             // Update state
-            this.setState({page_path: newPagePath, data_path : newDataPath}, this.updateData.bind(this));
+            this.setState({pagePath: newPagePath, dataPath : newDataPath}, this.updateData.bind(this));
         }
     }
 
@@ -114,7 +114,7 @@ class ModelProvider extends Component {
     getPagePath() {
         // 1. The model is hierarchical
         // 2. The page path is provided as a property
-        return (this.state && this.state.cq_model && this.state.cq_model[Constants.PATH_PROP]) || this.props && this.props.page_path || '';
+        return (this.state && this.state.cqModel && this.state.cqModel[Constants.PATH_PROP]) || this.props && this.props.pagePath || '';
     }
 
     /**
@@ -125,7 +125,7 @@ class ModelProvider extends Component {
      * @protected
      */
     isPageModel() {
-        return !!(this.state && this.state.cq_model && this.state.cq_model.hasOwnProperty(Constants.HIERARCHY_TYPE_PROP) && this.state.cq_model[Constants.HIERARCHY_TYPE_PROP] === HierarchyConstants.hierarchyType.page);
+        return !!(this.state && this.state.cqModel && this.state.cqModel.hasOwnProperty(Constants.HIERARCHY_TYPE_PROP) && this.state.cqModel[Constants.HIERARCHY_TYPE_PROP] === HierarchyConstants.hierarchyType.page);
     }
 
     /**
@@ -147,8 +147,8 @@ class ModelProvider extends Component {
         // a child page isn't a piece of content of the parent page
         if (this.isPageModel() && path) {
             childAttrs.cqPagePath = path;
-        } else {
-            childAttrs.cqDataPath = this.state.data_path;
+        } else if (this.state.dataPath) {
+            childAttrs.cqDataPath = this.state.dataPath;
         }
 
         Object.keys(childAttrs).forEach(attr => element.dataset[attr] = childAttrs[attr]);
@@ -172,17 +172,15 @@ class ModelProvider extends Component {
      * @protected
      */
     getData() {
-        return PageModelManager.getData({pagePath: this.getPagePath(), dataPath: this.state.data_path, forceReload: this.props.force_reload});
+        return PageModelManager.getData({pagePath: this.getPagePath(), dataPath: this.state.dataPath, forceReload: this.props.forceReload});
     }
 
     render() {
         if (!this.props.children || this.props.children.length < 1) {
             return null;
         }
-
         // List and clone the children to passing the data as properties
-        return Children.map(this.props.children, child =>
-            React.cloneElement(child, { ref: this.state.data_path, cq_model: this.state.cq_model, cq_model_page_path: this.state.page_path, cq_model_data_path: this.state.data_path }));
+        return React.cloneElement(this.props.children, { ref: this.state.dataPath, cqModel: this.state.cqModel, cqModelPagePath: this.state.pagePath, cqModelDataPath: this.state.dataPath });
     }
 }